Update configuration and improve event handling in MeiliSearch sync
This commit includes a modification of the configuration in 'config.yml', changing the host to 'localhost' among other changes. Additionally, the handling of database events in 'mysql.py' was improved by adding a condition to check if the event rows are not empty before handling them. This was done to prevent errors caused by attempts to process empty event rows.